fix: webpack 4 support
This change is necessary for extract-svg-sprite-webpack-plugin to be compatible with webpack 4. In that version of webpack, the second argument to NormalModule.source is the runtime template used by the target compilation. If not provided, webpack code which attempts to use the runtime will fail. For instance, the HarmonyCompatibilityDependency class calls the defineEsModuleFlagStatement method on the runtime during source dependency creation.